991220-S

+++ Common/Models/CompanyProductStatistics.cs
@@ -8,6 +8,6 @@
         public string U_DateTimeFa { get; set; }
         public double CompanyCode { get; set; }
         public string CompanyName { get; set; }
-        public double ProductCount { get; set; }
+        public int ProductCount { get; set; }
     }
 }
